+++ Lib/__init__.py
@@ -2,9 +2,3 @@
 ## Stelle -> File Controls on File Explorer on Python
 ## archeron -> bulk validator for assets
 ## himeko -> movie render queue functions
-
-
-import __lib_topaz__ as Topaz
-import __lib_stelle__ as Stelle
-import __lib_archeron__ as Archeron
-import __lib_himeko__ as Himeko
